requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
python #==3.9.15 
tqdm #==4.63.0
pytorch #==1.11.0
torchvision #==0.12.0
joblib #==1.1.0
numpy #==1.19.2
timm == 0.6.7
scipy #==1.9.3